backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
pydantic>=2.8
pyyaml>=6.0
transformers>=4.44
torch>=2.3
accelerate>=0.33
